fixes in zip create method

+++ src/main/java/util/Zip.java
@@ -64,13 +64,17 @@ public class Zip {
       final Path root = zipFileSystem.getPath("/");
 
       // iterate over the files we need to add
-      for (Path src : filenames) {
+      for (final Path src : filenames) {
+        if (!Files.exists(src)) {
+          log.info("[zip] " + src + " was skipped because it does not exists");
+          continue;
+        }
         // add a file to the zip file system
         if (!Files.isDirectory(src)) {
-          final Path dest = zipFileSystem.getPath(root.toString(), src.toString());
-          final Path parent = dest.getParent();
+          Path dest = zipFileSystem.getPath(root.toString(), src.getFileName().toString());
+          Path parent = dest.getParent();
           if (Files.notExists(parent)) {
-            log.fine("Creating directory " + parent);
+            log.fine("[zip] Creating directory: " + parent);
             Files.createDirectories(parent);
           }
           Files.copy(src, dest, StandardCopyOption.REPLACE_EXISTING);
@@ -80,7 +84,8 @@ public class Zip {
             @Override
             public FileVisitResult visitFile(Path file, BasicFileAttributes attrs)
                 throws IOException {
-              final Path dest = zipFileSystem.getPath(root.toString(), file.toString());
+              Path relativeFile = src.getFileName().resolve(src.relativize(file));
+              Path dest = zipFileSystem.getPath(root.toString(), relativeFile.toString());
               Files.copy(file, dest, StandardCopyOption.REPLACE_EXISTING);
               return FileVisitResult.CONTINUE;
             }
@@ -88,12 +93,19 @@ public class Zip {
             @Override
             public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
                 throws IOException {
-              final Path dirToCreate = zipFileSystem.getPath(root.toString(), dir.toString());
+              Path relativeDir = src.getFileName().resolve(src.relativize(dir));
+              Path dirToCreate = zipFileSystem.getPath(root.toString(), relativeDir.toString());
               if (Files.notExists(dirToCreate)) {
-                log.fine("Creating directory " + dirToCreate);
+                log.fine("[zip] Creating directory: " + dirToCreate);
                 Files.createDirectories(dirToCreate);
               }
               return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
+              log.info("[zip] Creation failed with: " + file + " due to " + exc.toString());
+              return super.visitFileFailed(file, exc);
             }
           });
         }
@@ -213,7 +225,7 @@ public class Zip {
 
   private static FileSystem createZipFileSystem(Path path, boolean create) throws IOException {
     // convert the filename to a URI
-    final URI uri = URI.create("jar:file:" + path.toUri().getPath());
+    final URI uri = URI.create("jar:file:" + path.toUri().getRawPath());
 
     final Map<String, String> env = new HashMap<>();
     if (create) {
